requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
onnx>=1.14.0
onnxruntime>=1.15.0
gradio>=3.50.0
pillow>=9.5.0
tqdm>=4.65.0
pyyaml>=6.0
onnxscript>=0.1.0